dynamic-chatbot/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
openai
schedule
beautifulsoup4
python-dotenv
tiktoken